refactor: 컴포넌트 폴더 상세 폴더인 Pages 폴더 추가 및 테스트 코드 import 경로 변경
